# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AKSIS Core API
  version: 1.13.0
  description: >-
    Integrated Phase 03-15 REST contract covering authentication and tenant
    context, schools/classes/students, academic periods and student history,
    card lifecycle, and secure device provisioning/runtime operations.
  x-aksis-phases:
    - phase: 3
      resources: [Auth, Schools, Classes, Students]
    - phase: 4
      resources: [Academic Years, Student History, Cards]
    - phase: 5
      resources: [Devices]
    - phase: 6
      resources: [Gate Attendance, Offline Sync]
    - phase: 7
      resources: [Admin Context, Attendance Read Model, Device Inventory]
    - phase: 8
      resources: [Parent Linking, Children, Parent Timeline]
    - phase: 9
      resources: [Waste Transactions, Ranking, Summary]
    - phase: 10
      resources: [Extracurriculars, Sessions, Members, Attendance]
    - phase: 11
      resources: [Library Visits, Offline Sync, Class Summary]
    - phase: 12
      resources: [LED Content, Gateway State, Heartbeat, Acknowledgement]
    - phase: 13
      resources: [Command Center, Today Aggregate]
    - phase: 14
      resources: [Card Writer Jobs, Station Lease, Read-back Verification]
    - phase: 15
      resources: [Reporting, Audit, Monitoring]
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
    schoolId: []
paths:
...
  /devices/register:
    post:
      tags: [Devices]
      operationId: registerDevice
      summary: Register a device and issue its initial credential once
      description: Uses human bearer/tenant authentication and requires `device.manage`.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeviceRegister' } } } }
      responses:
        '201': { description: Device and one-time plaintext credential, content: { application/json: { schema: { $ref: '#/components/schemas/DeviceRegistrationResponse' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /devices/heartbeat:
    post:
      tags: [Devices]
      operationId: deviceHeartbeat
      summary: Authenticate a device and record telemetry
      security: [{ deviceAuth: [], deviceId: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeviceHeartbeat' } } } }
      responses:
        '200': { description: Heartbeat accepted, content: { application/json: { schema: { $ref: '#/components/schemas/DeviceHeartbeatResponse' } } } }
        '401': { $ref: '#/components/responses/AuthInvalid' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /devices/config:
    get:
      tags: [Devices]
      operationId: getDeviceConfig
      summary: Authenticate a device and obtain only its tenant-scoped configuration
      security: [{ deviceAuth: [], deviceId: [] }]
      responses:
        '200': { description: Device configuration, content: { application/json: { schema: { $ref: '#/components/schemas/DeviceConfigResponse' } } } }
        '401': { $ref: '#/components/responses/AuthInvalid' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /devices:
    get:
      tags: [Devices]
      operationId: listDevices
      summary: List the selected school's device inventory
      description: Requires `device.read`; credential material is never returned.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
      responses:
        '200': { description: Paginated device inventory, content: { application/json: { schema: { $ref: '#/components/schemas/DeviceListResponse' } } } }
        '401': { $ref: '#/components/responses/AuthRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /platform/devices:
    get:
      tags: [Platform]
      operationId: platformDevices
      summary: "Paginated device status; online uses a two-minute heartbeat window"
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /platform/devices/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    patch:
      tags: [Platform]
      operationId: platformDeviceStatus
      summary: "Set administrative device status"
      security: [{ bearerAuth: [] }]
      requestBody: {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["status"], "properties": {"status": {"type": "string", "enum": ["ACTIVE", "DISABLED", "MAINTENANCE", "RETIRED"]}}}}}}
      responses: { '200': { description: Success envelope }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    schoolId:
      type: apiKey
      in: header
      name: X-School-Id
      description: UUID of the active school membership.
    deviceAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Per-device credential formatted as `Device <token>`.
    deviceId:
      type: apiKey
      in: header
      name: X-Device-Id
      description: Registered device UUID; school context is derived server-side.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Search:
      name: search
      in: query
      schema: { type: string, minLength: 1, maxLength: 100, pattern: '^[\p{L}\p{N}\s.-]+$' }
  responses:
...
    AuthRequired:
      description: Missing, expired, or invalid bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    AuthInvalid:
      description: Invalid session credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Forbidden:
      description: Tenant membership or permission is missing
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found in the active tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Resource violates a uniqueness constraint
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationError:
      description: Request payload, parameters, or query are invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, error, request_id]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [AUTH_REQUIRED, AUTH_INVALID, INVALID_CREDENTIALS, SESSION_REFRESH_FAILED, FORBIDDEN, TENANT_REQUIRED, SCHOOL_NOT_FOUND, ACADEMIC_YEAR_NOT_FOUND, CLASS_NOT_FOUND, STUDENT_NOT_FOUND, CARD_NOT_FOUND, DEVICE_NOT_REGISTERED, DEVICE_AUTH_INVALID, DEVICE_NOT_GATE, ATTENDANCE_DISABLED, CARD_BLOCKED, STUDENT_INACTIVE, GATE_RULE_VIOLATION, PARENT_LINK_INVALID, CHILD_NOT_LINKED, RESOURCE_NOT_FOUND, INVALID_JSON, PAYLOAD_TOO_LARGE, VALIDATION_ERROR, CONFLICT, RATE_LIMITED, ROUTE_NOT_FOUND, DATABASE_ERROR, SERVER_UNAVAILABLE, INTERNAL_ERROR]
            message: { type: string }
            details: {}
        request_id: { type: string }
    PaginationMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...
    DeviceType: { type: string, enum: [GATE, LIBRARY, LED, CARD_STATION, WASTE_SCALE, OTHER] }
    DeviceRegister:
      type: object
      additionalProperties: false
      required: [device_code, device_type, name]
      properties:
        device_code: { type: string, minLength: 1, maxLength: 64 }
        device_type: { $ref: '#/components/schemas/DeviceType' }
        name: { type: string, minLength: 1, maxLength: 120 }
        location: { type: [string, 'null'], maxLength: 200 }
        hardware_version: { type: [string, 'null'], maxLength: 50 }
        firmware_version: { type: [string, 'null'], maxLength: 50 }
        update_channel: { type: string, default: stable, maxLength: 32 }
        config: { type: object, additionalProperties: true }
    DeviceHeartbeat:
      type: object
      additionalProperties: false
      required: [uptime_seconds, firmware_version, reported_at]
      properties:
        uptime_seconds: { type: integer, minimum: 0 }
        signal_strength: { type: [integer, 'null'], minimum: -150, maximum: 0 }
        firmware_version: { type: string, minLength: 1, maxLength: 50 }
        storage_status: { type: object, additionalProperties: true }
        last_error: { type: [string, 'null'], maxLength: 1000 }
        reported_at: { type: string, format: date-time }
    DeviceRegistrationResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                device: { type: object, additionalProperties: true }
                credential:
                  type: object
                  required: [token, scheme]
                  properties:
                    token: { type: string, description: Returned once; store in device secure storage. }
                    scheme: { type: string, const: Device }
    DeviceHeartbeatResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [accepted, received_at]
              properties:
                accepted: { type: boolean, const: true }
                received_at: { type: string, format: date-time }
    DeviceConfigResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [device_id, school_id, device_code, device_type, config]
              properties:
                device_id: { type: string, format: uuid }
                school_id: { type: string, format: uuid }
                device_code: { type: string }
                device_type: { $ref: '#/components/schemas/DeviceType' }
                config: { type: object, additionalProperties: true }
                update_channel: { type: string }
                firmware_version: { type: [string, 'null'] }
...
    DeviceInventoryItem:
      type: object
      required: [id, school_id, device_code, device_type, name, status]
      properties:
        id: { type: string, format: uuid }
        school_id: { type: string, format: uuid }
        device_code: { type: string }
        device_type: { $ref: '#/components/schemas/DeviceType' }
        name: { type: string }
        location: { type: [string, 'null'] }
        firmware_version: { type: [string, 'null'] }
        hardware_version: { type: [string, 'null'] }
        update_channel: { type: string }
        status: { type: string, enum: [PROVISIONING, ACTIVE, DISABLED, MAINTENANCE, RETIRED] }
        last_seen_at: { type: [string, 'null'], format: date-time }
    DeviceListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/DeviceInventoryItem' } }
            meta: { $ref: '#/components/schemas/PaginationMeta' }
...